remove testutil pkg
